fix(web-client): address validator for bookmark.
